Ability Tutorial Update (#47)
* Update abilities_advanced.md

Fixed some broken indentation in ability phase explanations, and changed headshot section to be udpated for targetInteractionEvent parameter changes.

* Update abilities.md

first teeeny update before lunch on this tutorial--just needs images and text changes from dance to wave.

* First Ability Tutorial changed to Wave

Updated the ability tutorial to be an animation that exists again. Replaced all images and references to the old animation.

This included updating the index page to display the proper preview image.

* Update abilities.md

first teeeny update before lunch on this tutorial--just needs images and text changes from dance to wave.

* First Ability Tutorial changed to Wave

Updated the ability tutorial to be an animation that exists again. Replaced all images and references to the old animation.

This included updating the index page to display the proper preview image.

* Optimised images with calibre/image-actions

* Updated index.md

Changed ability tutorial preview image to an mp4 instead of a gif.