Wrote more thorough outline
Laid out the first pass of an overview section, as well as the outline for the tutorial, and started a tips section. Included Jishnu's notes for my own reference, as well as a placeholder image to visualize space breakup. Ready to start implementing real info once I recieve the examples from Tung & Jishnu.