Finished tutorial
Finished a polish pass and replaced images. This could in total be nicer and more thorough, but for the release next week this explains all the necessary parts in still an easy-enough way for anyone to understand. Took out the examples section because there won't be any until the update is live on prod.