Persistent Storage Tutorial (#76)
* Started outline for persistence tutorial

Created a page, gutted it and replaced some basic words; setting up for getting into the writing portion. Added to directories.

* Wrote more thorough outline

Laid out the first pass of an overview section, as well as the outline for the tutorial, and started a tips section. Included Jishnu's notes for my own reference, as well as a placeholder image to visualize space breakup. Ready to start implementing real info once I recieve the examples from Tung & Jishnu.

* Update persistent_storage.md

Finished the loose version of all steps, tutorial can now be completed in bare-bones form.

* Cleaning up tutorial

Ran through this again to add more clarity and temporary images/gifs (which will be videos later) not quite done, but it is fully completeably and pretty polished. Added code explanations alongside code.

* Finished tutorial

Finished a polish pass and replaced images. This could in total be nicer and more thorough, but for the release next week this explains all the necessary parts in still an easy-enough way for anyone to understand. Took out the examples section because there won't be any until the update is live on prod.

* Optimised images with calibre/image-actions

* fix formatting

* spelling fixes

Co-authored-by: Benjamin Staneck <bstaneck@manticoregames.com>