add lens flare section
Finished up/polished the AO recolor section, and layed out the lensflare section. Techincally finished that part, but wording/steps will probably be improved. This concludes the PPE covered by the video, so I will next be moving onto the stacking of vfx and triggerable fx. Then, I hope to move back to PPE as there are some amazing ones that are still not mentioned.