VFX tutorial (#145)
* Create vfx_tutorial.md

Started blocking out a page for the VFX tutorial, setting up what to write and spending time in the VFX demo level to choose what sort of thing to be made. Also going to include a section that explains going through the demo level, as these should definitely go together.

* Added to Overview & Improved Outline

Cleaned up the overview & explanations of locations for vfx assets; further preparing for fleshing out the two halves of the tutorial. Added the page to the directory.

* Finished cleanup of VFX tutorial page

Cleaned up overview and reference section, once the video exists we can just plug it into the bottom of the page. Added some images explaining where to find different VFX.

* starting PPE tutorial section

Following the topics in the video series, I've started laying out the written tutorial.  Changed all CORE to Core and included more overall sections about PPE, and the beginnings of a tutorial on advanced bloom. Also changed the mkdocs title for more clarity.

* More VFX updates

Removed hyphen from some parts of post processing to be consistent. Finished the section explaining Advanced Bloom, and am moving onto AO. Reorganized structure of tutorial parts so they are more distinct.

* ambient occlusion section addition

Finished up the section on ambient occlusion, and moved onto the recolor portion of that. Added more images to go with this.

* AO Recolor & Lens Flares

Finished more explanation of the AO recolor PPE, and started the lens flare section.

* add lens flare section

Finished up/polished the AO recolor section, and layed out the lensflare section. Techincally finished that part, but wording/steps will probably be improved. This concludes the PPE covered by the video, so I will next be moving onto the stacking of vfx and triggerable fx. Then, I hope to move back to PPE as there are some amazing ones that are still not mentioned.

* finish lensflare section

Finished the lensflare section and thus the whole post process effects section. Also added credits to creators whose CC templates are used.

* Update vfx_tutorial.md

Added numbered steps to more parts of the post process section as it had started to become more of a reference/explanation.

Laid out loose ideas for how the interaction section should go. Spent time in Core setting up what that would be, and watching the heartbeast video to see what is done in that.

* Adding VFX tutorial steps

started a rough outline of moment-to-moment steps from the treasure chest vfx video that will be fleshed out next.

* Update vfx_tutorial.md

Transcribed more steps from the vfx tutorial video in prep for fleshing out later. This tutorial is now on hold while I try to update as many existing tutorials as possible.

* Update vfx_tutorial.md

Continuing writeup of the VFX tutorial. Finished outlining steps from Heartbeast video, and started fleshing out the steps from start to finish more solidly.

* Update vfx_tutorial.md

Fleshed out more steps of the vfx tutorial section.

* Update vfx_tutorial.md

Wrote more steps & added broken-up sections so that it can be done in clear parts.

* Update vfx_tutorial.md

added some script previews, pushing before rebase (because I have to)

* Update vfx_tutorial.md

Finished more sections of the tutorial, only part of the last one left.

* Update vfx_tutorial.md

a couple more steps to the vfx tutorial written out--also pushing to confirm everything is working.

* finished vfx tutorial

Fnished writing all the tutorial steps and adding images. Needs a proofread, and two videos are still gifs that need to be converted. Otherwise, fully completeable!

* Replaced remaining gifs with mp4's

Replaced the last two gifs with mp4 files. Now it's ready!

* Update vfx_tutorial.md

Fixed broken links.

* Fix conflicts

* Optimised images with calibre/image-actions

* Crush images

* Optimised images with calibre/image-actions

* Replace every gif with a video

* add webm versions for every video

* CORE -> Core

* Add videos

Co-authored-by: Sierra Jones <rhiktertorn@gmail.com>
Co-authored-by: Benjamin Staneck <bstaneck@manticoregames.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

Former-commit-id: 4a85527a0c877bdf9c0bf5800685852318751f53