Loop added to video tag.
Since these are quite small it should be ok to loop these to be consistent with the rest of the docs.  Originally I removed the attribute but in doing so videos would reach their end point while not in the viewport and it may not be obvious to the use it's a video rather than a static image.