Corrected: Function written in camel case
In addition to changing the function to Pascal Case, I've never seen "local" in front of "function" either and am assuming it's most probably a typo, so it was removed. If there is a reason this function is camel-case, it should probably be explained, as this will confuse the heck out of new users. I'm confused and I'm working on two years in. The Lua Style Guide at Core says functions are Pascal Case.