Fixed: "Activate() function is client-only"
I don't think this function is client-only. I've been using it for months in a default context, albeit Networked. If it is client-only, that should be listed as a tag in the Functions section (currently, it's "None"). As described later in the paragraph, using Activate() sends a broadcast to the client. Does this statement also need amended, or is it accurate? Further explanation of what happens next would be good so creators can understand the process, as Abilities are one of the most complex Core Components to work with for a variety of reasons.